+++ zoomimage-core-glide/src/main/kotlin/com/github/panpf/zoomimage/glide/GlideModelToImageSource.kt
@@ -10,39 +10,40 @@ import com.github.panpf.zoomimage.subsampling.fromByteArray
 import com.github.panpf.zoomimage.subsampling.fromContent
 import com.github.panpf.zoomimage.subsampling.fromFile
 import com.github.panpf.zoomimage.subsampling.fromResource
+import com.github.panpf.zoomimage.subsampling.toFactory
 import java.io.File
 import java.net.URL
 
 interface GlideModelToImageSource {
-    fun dataToImageSource(model: Any): ImageSource?
+    fun dataToImageSource(model: Any): ImageSource.Factory?
 }
 
 class GlideModelToImageSourceImpl(private val context: Context) : GlideModelToImageSource {
 
-    override fun dataToImageSource(model: Any): ImageSource? {
+    override fun dataToImageSource(model: Any): ImageSource.Factory? {
         return when {
             model is GlideUrl -> {
-                GlideHttpImageSource(Glide.get(context), model)
+                GlideHttpImageSource.Factory(Glide.get(context), model)
             }
 
             model is URL -> {
-                GlideHttpImageSource(Glide.get(context), GlideUrl(model))
+                GlideHttpImageSource.Factory(Glide.get(context), GlideUrl(model))
             }
 
             model is String && (model.startsWith("http://") || model.startsWith("https://")) -> {
-                GlideHttpImageSource(Glide.get(context), model)
+                GlideHttpImageSource.Factory(Glide.get(context), model)
             }
 
             model is Uri && (model.scheme == "http" || model.scheme == "https") -> {
-                GlideHttpImageSource(Glide.get(context), model.toString())
+                GlideHttpImageSource.Factory(Glide.get(context), model.toString())
             }
 
             model is String && model.startsWith("content://") -> {
-                ImageSource.fromContent(context, Uri.parse(model))
+                ImageSource.fromContent(context, Uri.parse(model)).toFactory()
             }
 
             model is Uri && model.scheme == "content" -> {
-                ImageSource.fromContent(context, model)
+                ImageSource.fromContent(context, model).toFactory()
             }
 
             model is String && model.startsWith("file:///android_asset/") -> {
@@ -50,7 +51,7 @@ class GlideModelToImageSourceImpl(private val context: Context) : GlideModelToIm
                     .takeIf { it.size > 1 }
                     ?.let { it.subList(1, it.size) }
                     ?.joinToString(separator = "/")
-                assetFileName?.let { ImageSource.fromAsset(context, it) }
+                assetFileName?.let { ImageSource.fromAsset(context, it).toFactory() }
             }
 
             model is Uri && model.scheme == "file" && model.pathSegments.firstOrNull() == "android_asset" -> {
@@ -58,29 +59,29 @@ class GlideModelToImageSourceImpl(private val context: Context) : GlideModelToIm
                     .takeIf { it.size > 1 }
                     ?.let { it.subList(1, it.size) }
                     ?.joinToString(separator = "/")
-                assetFileName?.let { ImageSource.fromAsset(context, it) }
+                assetFileName?.let { ImageSource.fromAsset(context, it).toFactory() }
             }
 
             model is String && model.startsWith("file://") -> {
                 val filePath = Uri.parse(model).path
-                filePath?.let { ImageSource.fromFile(File(filePath)) }
+                filePath?.let { ImageSource.fromFile(File(filePath)).toFactory() }
             }
 
             model is Uri && model.scheme == "file" -> {
                 val filePath = model.path
-                filePath?.let { ImageSource.fromFile(File(filePath)) }
+                filePath?.let { ImageSource.fromFile(File(filePath)).toFactory() }
             }
 
             model is File -> {
-                ImageSource.fromFile(model)
+                ImageSource.fromFile(model).toFactory()
             }
 
             model is Int -> {
-                ImageSource.fromResource(context, model)
+                ImageSource.fromResource(context, model).toFactory()
             }
 
             model is ByteArray -> {
-                ImageSource.fromByteArray(model)
+                ImageSource.fromByteArray(model).toFactory()
             }
 
             else -> {