+++ zoomimage-core-sketch3/src/main/kotlin/com/github/panpf/zoomimage/sketch/SketchImageSource.kt
@@ -63,10 +63,9 @@ class SketchImageSource(
         override val key: String = imageUri
 
         override suspend fun create(): SketchImageSource {
-            // TODO support download
             val request = LoadRequest(context, imageUri) {
                 downloadCachePolicy(CachePolicy.ENABLED)
-                depth(Depth.LOCAL)   // Do not download image, by default go here The image have been downloaded
+                depth(Depth.NETWORK)
             }
             val fetcher = sketch.components.newFetcherOrThrow(request)
             val fetchResult = fetcher.fetch().getOrThrow()
@@ -94,7 +93,7 @@ class SketchImageSource(
         }
 
         override fun toString(): String {
-            return "SketchImageSource('$imageUri')"
+            return "SketchImageSource.Factory('$imageUri')"
         }
     }
 }